kyc-extraction/requirements.txt
python-dotenv
dnspython
mongoengine
pymongo
PyYAML
uvicorn
fastapi
openai
langchain
pydantic
redis
websockets
rq
bcrypt
PyJWT
requests
JinJa2
httpx
stripe
certifi
black
boto3
python-multipart
celery
watchdog